New Development Mode!! (WIP)
The new dev mode is almost here, you can try it with:
   $ yarn leemons newdev

The leemons app is now splitted in 3 main parts:
 - Config
 - Front
 - Back

So leemons can be restarted partially gaining development time,
when a frontend or backend file is updated, the modified service
is reloaded, so the changes take effect, this is possible because
they are independent.

In the case of the Config service, it will need to reload the whole
app, because the Front and the Back needs this updated config, so
the entire app is taken down and then raised up with the new config.

How do we know which service has been modified?

Well, each service uses some files to work:

 - CONFIG:
  The config service uses all the leemons-core packages (all those
  that starts with "leemons-" but not "leemons-plugin" or
  "leemons-connector" and the app config (leemons.dir.config).

 - FRONT:
  The app' next.js directory (leemons.dir.next) and each plugin next
  folder (plugin.dir.next).

 - BACK:
  Each plugin models, controllers and services folder
  (plugin.dir.models, plugin.dir.controllers, plugin.dir.services)

The reload in the frontend is made in two steps, the first one is
checking and copying all the frontend related files and the second
one is using the next.js development reload (this way, we can avoid
the long compilation times, and see no-obfuscated names in
react-dev-tools).

The reload in the backend is also made in two steps, the first one is
removing all the backend endpoints from the BackendRouter (for freeing
the old files) and the second one is calling `leemons.buildBack` for
initializing all the backend stuff.