feat(common): Add useProcessTextEditor
Hook for processing TextEditor output, it returns an async function, this one, accepts 2 arguments, the first one, the current html (TextEditor output), the second, is the last saved html from TextEditor (This argument is used to remove unused data from database).

feat(leebrary): Register libraryProcessor on useProcessTextEditor
Add a new processor to parse the html in search of assets and replace them with its duplicated version (to be public and non-indexable), so it can be seen by all the users able to see the output.
If the asset is already duplicated, it will be marked as processed, so it's not duplicated again.

feat(leebrary): Add processed prop to the ASSET_KEYS in LibraryExtension

fix(leebrary): Return undefined url at appendAuthorizationToUrl when the url is null or "null"

feat(leebrary): Expose the preserveName, public and indexable to API calls on duplicateAsset

feat(tasks): Process the development output through the new processTextEditor

fix(tasks): Set dependency to an array at Curriculum component