feat(users): allow admins to impersonate users
When impersonating any user, the admin can act in behalf of the impersonated user. This means that it can view and update anything the user has access to. The actions made as the impersonated user will be recorded as made by the user and not the admin, but the impersonation will be recorded in xAPI.
Ticket: task/Users-153
Reviewed-by: @MIGUELez11
Refs: #611